+++ packages/core/tests/reporters/CIReporter.test.ts
@@ -1,5 +1,5 @@
 import chalk from 'chalk';
-import { mockTool, mockConsole } from '../../src/testUtils';
+import { mockTool, mockConsole, mockRoutine } from '../../src/testUtils';
 import CIReporter from '../../src/reporters/CIReporter';
 
 describe('CIReporter', () => {
@@ -20,16 +20,15 @@ describe('CIReporter', () => {
 
   describe('bootstrap()', () => {
     it('binds events', () => {
-      const spy = jest.spyOn(reporter.console, 'on');
+      const startSpy = jest.spyOn(reporter.console.onStart, 'listen');
+      const routineSpy = jest.spyOn(reporter.console.onRoutine, 'listen');
+      const taskSpy = jest.spyOn(reporter.console.onTask, 'listen');
 
       reporter.bootstrap();
 
-      expect(spy).toHaveBeenCalledWith('stop', expect.anything());
-      expect(spy).toHaveBeenCalledWith('task', expect.anything());
-      expect(spy).toHaveBeenCalledWith('routine', expect.anything());
-      expect(spy).toHaveBeenCalledWith('routine.skip', expect.anything());
-      expect(spy).toHaveBeenCalledWith('routine.pass', expect.anything());
-      expect(spy).toHaveBeenCalledWith('routine.fail', expect.anything());
+      expect(startSpy).toHaveBeenCalledWith(expect.anything());
+      expect(routineSpy).toHaveBeenCalledWith(expect.anything());
+      expect(taskSpy).toHaveBeenCalledWith(expect.anything());
     });
   });
 
@@ -53,13 +52,13 @@ describe('CIReporter', () => {
     it('increments count', () => {
       expect(reporter.routineCount).toBe(0);
 
-      reporter.handleRoutine();
+      reporter.handleRoutine(mockRoutine(reporter.tool));
 
       expect(reporter.routineCount).toBe(1);
     });
 
     it('logs a period', () => {
-      reporter.handleRoutine();
+      reporter.handleRoutine(mockRoutine(reporter.tool));
 
       expect(outSpy).toHaveBeenCalledWith('.');
     });