+++ packages/event/src/constants.ts
@@ -1,0 +1,3 @@
+export const DEFAULT_SCOPE = '*';
+
+export const EVENT_NAME_PATTERN = /^[a-z]{1}[-.a-z0-9]*[a-z]{1}$/u;