+++ packages/reporter-nyan/tests/NyanReporter.test.ts
@@ -16,16 +16,17 @@ describe('NyanReporter', () => {
 
   describe('bootstrap()', () => {
     it('binds events', () => {
-      const spy = jest.spyOn(reporter.console, 'on');
+      const startSpy = jest.spyOn(reporter.console.onStart, 'listen');
+      const stopSpy = jest.spyOn(reporter.console.onStop, 'listen');
+      const routineSpy = jest.spyOn(reporter.console.onRoutine, 'listen');
+      const taskSpy = jest.spyOn(reporter.console.onTask, 'listen');
 
       reporter.bootstrap();
 
-      expect(spy).toHaveBeenCalledWith('start', expect.anything());
-      expect(spy).toHaveBeenCalledWith('stop', expect.anything());
-      expect(spy).toHaveBeenCalledWith('routine', expect.anything());
-      expect(spy).toHaveBeenCalledWith('routine.pass', expect.anything());
-      expect(spy).toHaveBeenCalledWith('routine.fail', expect.anything());
-      expect(spy).toHaveBeenCalledWith('task', expect.anything());
+      expect(startSpy).toHaveBeenCalledWith(expect.anything());
+      expect(stopSpy).toHaveBeenCalledWith(expect.anything());
+      expect(routineSpy).toHaveBeenCalledWith(expect.anything());
+      expect(taskSpy).toHaveBeenCalledWith(expect.anything());
     });
 
     it('generates rainbow data', () => {